Separate code transformation logic from source code retrieval
So that it could be synchroneously used in the workers
